Move AvaloniaWindowUtils to special folder
